add home school UI refactor small classroom to purchase component for re-use
